requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
Flask-Login>=0.6
Flask-Dance>=7.1
SQLAlchemy>=2.0
psycopg2-binary>=2.9
Werkzeug>=3.0
python-dotenv>=1.0
requests>=2.31
gunicorn>=21.2
langchain-core>=0.3
langchain-groq>=0.2
scikit-learn>=1.4
joblib>=1.3
numpy>=1.26
